requirements.txt
streamlit>=1.36
sqlalchemy>=2.0
psycopg[binary]>=3.2
pgvector>=0.3.2
pypdf>=5.0
openai>=1.40
python-dotenv>=1.0
sentence-transformers>=4.1.0
pytorch>=1.0.2